Move Attributes module into its own file
