+++ lib/json_schema.rb
@@ -1,3 +1,4 @@
+require_relative "json_schema/attributes"
 require_relative "json_schema/configuration"
 require_relative "json_schema/document_store"
 require_relative "json_schema/error"
