+++ lib/json_schema/schema.rb
@@ -1,75 +1,6 @@
 require "json"
 
 module JsonSchema
-  module Attributes
-    module ClassMethods
-      attr_reader :copyable_attrs
-
-      # Attributes that are part of the JSON schema and hyper-schema
-      # specifications. These are allowed to be accessed with the [] operator.
-      #
-      # Hash contains the access key mapped to the name of the method that should
-      # be invoked to retrieve a value. For example, `type` maps to `type` and
-      # `additionalItems` maps to `additional_items`.
-      attr_reader :schema_attrs
-
-      # identical to attr_accessible, but allows us to copy in values from a
-      # target schema to help preserve our hierarchy during reference expansion
-      def attr_copyable(attr)
-        attr_accessor(attr)
-        @copyable_attrs ||= []
-        self.copyable_attrs << "@#{attr}".to_sym
-      end
-
-      def attr_schema(attr, options = {})
-        attr_copyable(attr)
-        @schema_attrs ||= {}
-        self.schema_attrs[options[:schema_name] || attr] = attr
-      end
-
-      def attr_reader_default(attr, default)
-        # remove the reader already created by attr_accessor
-        remove_method(attr)
-
-        class_eval("def #{attr} ; !@#{attr}.nil? ? @#{attr} : #{default} ; end")
-      end
-    end
-
-    def self.included(klass)
-      klass.extend(ClassMethods)
-    end
-
-    # Allows the values of schema attributes to be accessed with a symbol or a
-    # string. So for example, the value of `schema.additional_items` could be
-    # procured with `schema[:additionalItems]`. This only works for attributes
-    # that are part of the JSON schema specification; other methods on the
-    # class are not available (e.g. `expanded`.)
-    #
-    # This is implemented so that `JsonPointer::Evaluator` can evaluate a
-    # reference on an sintance of this class (as well as plain JSON data).
-    def [](name)
-      name = name.to_sym
-      if self.class.schema_attrs.key?(name)
-        send(self.class.schema_attrs[name])
-      else
-        raise NoMethodError, "Schema does not respond to ##{name}"
-      end
-    end
-
-    def copy_from(schema)
-      self.class.copyable_attrs.each do |copyable|
-        instance_variable_set(copyable, schema.instance_variable_get(copyable))
-      end
-    end
-
-
-    def initialize_schema_attrs
-      self.class.schema_attrs.each do |_, a|
-        send(:"#{a}=", nil)
-      end
-    end
-  end
-
   class Schema
     include Attributes
 
@@ -82,6 +13,8 @@ module JsonSchema
       # variables.
       initialize_schema_attrs
     end
+
+    initialize_attrs
 
     # Fragment of a JSON Pointer that can help us build a pointer back to this
     # schema for debugging.
@@ -349,7 +282,9 @@ module JsonSchema
     end
 
     # Link subobject for a hyperschema.
-    Link = Schema
+    class Link < Schema
+      inherit_attrs
+    end
 
     # Media type subobject for a hyperschema.
     class Media
